refactor: input 폴더로 분리, 사용성 증대
